langchain_agent/requirements.txt
fastapi
uvicorn
python-dotenv
requests
langchain
langchain-core
langchain-community
langchain-ollama
SQLAlchemy
